+++ src/mlframe/metrics/calibration/_calibration_metrics.py
@@ -22,7 +22,7 @@ from math import floor
 import numpy as np
 import numba
 
-from ._numba_params import NUMBA_NJIT_PARAMS
+from .._numba_params import NUMBA_NJIT_PARAMS
 from ._calibration_plot import fast_calibration_binning